+++ src/main/java/pl/umcs/workshop/game/GameRepository.java
@@ -2,6 +2,6 @@ package pl.umcs.workshop.game;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameRepository extends JpaRepository<Game, Integer> {
+public interface GameRepository extends JpaRepository<Game, Long> {
 
 }
