+++ src/main/java/pl/umcs/workshop/round/RoundRepository.java
@@ -2,11 +2,11 @@ package pl.umcs.workshop.round;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoundRepository extends JpaRepository<Round, Integer> {
+public interface RoundRepository extends JpaRepository<Round, Long> {
 
-    Round findOneByGameIdAndUserOneIdAndUserTwoIdAndGeneration(int gameId, int userOneId, int userTwoId, int generation);
+    Round findOneByGameIdAndUserOneIdAndUserTwoIdAndGeneration(Long gameId, Long userOneId, Long userTwoId, int generation);
 
-    default Round getNextRound(int gameId, int userId, int generation) {
+    default Round getNextRound(Long gameId, Long userId, int generation) {
         return findOneByGameIdAndUserOneIdAndUserTwoIdAndGeneration(gameId, userId, userId, generation);
     }
 }
