+++ src/main/java/pl/umcs/workshop/topology/TopologyRepository.java
@@ -2,5 +2,5 @@ package pl.umcs.workshop.topology;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TopologyRepository extends JpaRepository<Topology, Integer> {
+public interface TopologyRepository extends JpaRepository<Topology, Long> {
 }
