+++ src/main/java/pl/umcs/workshop/user/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-    List<User> findAllByGameId(int gameId);
+public interface UserRepository extends JpaRepository<User, Long> {
+    List<User> findAllByGameId(Long gameId);
 }
